fix/move joinroles to own command
